Added obligatory arrows
In order to economise on the number of switches we need in a decision flow, it's handy
to support arrows that MUST be executed regardless of the expression computed in the origin switchn.

This particularly applies to "don't cares" found in the Snowball decision table.

So, after this commit, an arrow may of one of the following types:

- ORDINARY:   will be followed if the arrow expression computes to the same value as the
origin switch expression
- DEFAULT:    will be executed if no matching ordinary arrow is found
- OBLIGATORY: will be always executed