Added skipLoopDetection to DecisionFlow
The usage of the arrows of type <<always>> in the current version of the decision table may lead
to loops in the graph, so, if we still want to use them, we have to be able to relax this check.

In addition, pretty prints of decision paths are made even prettier.